move popover to separate component
